requirements.txt
fastapi
uvicorn==0.29.0
torch==2.2.2
ultralytics==8.1.37
deepface==0.0.93
huggingface_hub==0.23.1
pillow==9.5.0
opencv-python-headless==4.8.1.78
matplotlib==3.7.1
sentence-transformers==2.6.1
geopy==2.4.1
langdetect==1.0.9
deep-translator==1.11.4

